Option to avoid flattening arrays of primitive types
Normally,

  {"a": [1, 2, 3], "m": {"key": "value"}}

would be flattened to,

  {"a.0": 1, "a.1": 2, "a.2": 3, "m.key": "value"}

But in many cases, it's preferable to leave arrays of primitive types (such
as numbers and strings) as they are, e.g.

  ("a": [1, 2, 3], "m.key": "value"}

The new SeparatorStyle.DoNotFlattenPrimitiveArrays option is off by default
so the default behaviour doesn't change. Sadly, that's also why it has to be
an ugly negative option.